Demonstrate that incorrect password doesn't result in login success.
This is really more up to the implementor of Persona than it is up to
Persona proper, but I thought fleshing out the test suite for this
helper library would be a good thing.  Also, this forces me to come up
with a good way to clear Persona state between tests.