Implement a way to click the "this is not me" link.
This commit also removes memoization of the Persona window.  When
clicking the "this is not me" link, it's quite possible that that link
will be in a window with a different identifier.

It'd be useful to have a way to see if that button was visible, too.
Not sure of the best way to pull that off.