Short-circuit host resolution.
When given an email of the form foo@example.org, the BrowserID code will
attempt to check if browserid.org is an IdP.  If it isn't, the fallback
IdP will be used.

For some reason (I haven't yet figured out why) this process times out
on Travis' workers.  This means that the GET /.well-known/browserid
takes a very long time; this in turn causes the BrowserID UI code to
stall, which in turn causes timeouts to show up in our scenarios.

(Ugh.)

Luckily, we don't *need* to contact example.org for anything, because we
don't care whether or not a fallback is involved.  We're not testing
BrowserID; we're assuming it works, and we're testing whether or not our
driver code works.  Additionally, example.org should be considered a
reserved second-level domain for testing purposes (see RFC 2606), so I'm
fine with disconnecting it from the Internet, which is what this commit
(more or less) does.